[Breaking] Migrate from jqueryify2 to jqueryify
The [jqueryify2](https://npmjs.org/package/jqueryify2) project was forked from
[jqueryify](https://npmjs.org/package/jqueryify) due to a slight incompatibility
in headless tests. Because of this, Quips was locked to an older version of
jQuery.

In the passing time, the jqueryify library has integrated a fix and migrated to
a more recent jQuery library (1.9.1). This library has a number of small
incompatibilities with Quips' expectations, outlined below.

* Empty response bodies with a 200 status are treated as failures. The correct
  status should be 204. Rather than checking for this behavior and correct the
  HTTP response status in Quips, I'm opting to require consumers fix this, as an
  empty response is distinctly different than one with a body.
* The jQuery.Deferred.is* methods appear to have been replaced with a .state()
  method which returns a string.
* select elements on the DOM now treat .val() like a text field, however, the
  DOM does not update the option element with a "selected" property.