VBID-2214: fix how PageableCollection paginates FilteredCollection
Before, if you added a filter to a FilteredCollection that removed multiple items
from the current page, the page would end up containing fewer items than the given
page size. Ensuring that we always grab a new item for the page when one is removed
and that that item is the next one not on the page fixes this.